PLAT-301 Moved Pipe struct and load func to config
